rename app/get_config to app/config
